Testcase 단일화, 객체별 상속관계 수정, Monster run()메서드 추가.
한번의 테스트 실행을 통해, 3가지 객체가 제대로 작동하는지 검사.
interface가 추상 객체인 Monster를 통해서 구현되도록 변경. 좀 더 단조로운 상속관계를 형성하기 위함.
IsForward()가 재귀적으로 작동하도록 수정함.
게임이 작동될 Main Class 생성
