--- src/main/java/action/Run.java
@@ -1,16 +1,0 @@
-package action;
-
-import java.util.Random;
-import java.util.Stack;
-
-public class Run {
-    private Random random;
-    private Stack<Integer> isForwardStack;
-
-    public Run() {
-        this.random = new Random();
-        this.isForwardStack = new Stack<>();
-    }
-
-
-}
