--- src/test/java/FlyingTest.java
@@ -1,13 +1,0 @@
-import domain.Flying;
-import org.junit.Test;
-
-public class FlyingTest {
-
-    @Test
-    public void createFlying(){
-        Flying flying = new Flying("포비");
-        flying.attempt(10);
-        flying.isForward();
-        flying.forward();
-    }
-}
